[cm] Split up scattering and make jtfst methods
